api/requirements.txt
# SATH Platform API - Dependencies
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0

# Database
asyncpg==0.29.0
sqlalchemy==2.0.25
alembic==1.13.1

# Auth
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Redis
redis==5.0.1

# Docker
docker==7.0.0

# Encryption
cryptography==42.0.0

# Billing
stripe==7.12.0

# Utils
httpx==0.26.0
pyyaml==6.0.1
python-dotenv==1.0.0
